clean up and added more unit tests
