+++ packages/transformers-llguidance/src/utils/runtime.ts
@@ -1,0 +1,3 @@
+export function isNodeLikeRuntime() {
+    return typeof process !== 'undefined' && Boolean(process.versions?.node);
+}
